+++ CHANGELOG.md
@@ -16,8 +16,13 @@
   * All the functions which are applicable to data structures are moved into
     these structures, for example `wrapTable` is moved into `Table.wrap()`
   * All the named wrap functions are moved into `Janet` with a signature
-    `wrap(comptime T: type, value: T)` so you can use it like `Janet.wrap(i32, 3)`
+    `wrap(comptime T: type, value: T) Janet` so you can use it like
+    `Janet.wrap(i32, 3)`
   * `wrapNumberSafe` moved into `Janet.numberSafe`
+* `unwrap` function overhaul:
+  * All the functions except for abstract have changed their signature to
+    `unwrap(janet: Janet, comptime T: type) !T` so you can use it like
+    `try Janet.unwrap(i32)`
 * `string` function and corresponding for keyword and symbol are moved into
   their data structures into functions `init`
 * `stringv` function and corresponding for keyword and symbol are moved into